% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{create_shiny_app}
\alias{create_shiny_app}
\title{Create Shiny App}
\usage{
create_shiny_app()
}
\value{
Logical indicating success
}
\description{
Helper function to create Shiny app if it doesn't exist
}
