% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{detect_asset_classes}
\alias{detect_asset_classes}
\title{Detect Asset Classes}
\usage{
detect_asset_classes(data, asset_names = NULL)
}
\arguments{
\item{data}{Numeric matrix of returns}

\item{asset_names}{Character vector of asset names}
}
\value{
Character vector of detected asset classes
}
\description{
Automatically detect asset classes from data characteristics
}
