% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{get_welcome_banner}
\alias{get_welcome_banner}
\title{Get Welcome Banner}
\usage{
get_welcome_banner()
}
\value{
Character string with formatted banner
}
\description{
Generate the AgenticWaves welcome banner.
}
