% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{show_interactive_menu}
\alias{show_interactive_menu}
\title{Show Interactive Menu}
\usage{
show_interactive_menu()
}
\value{
Character string with user choice
}
\description{
Display the interactive menu options.
}
