% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{get_sample_data}
\alias{get_sample_data}
\title{Get Sample Data}
\usage{
get_sample_data(data_type = "global_markets", n_assets = 10, n_periods = 1000)
}
\arguments{
\item{data_type}{Character string specifying data type: "global_markets", "crypto", "commodities"}

\item{n_assets}{Integer number of assets to include (default: 10)}

\item{n_periods}{Integer number of time periods (default: 1000)}
}
\value{
Matrix of financial returns data
}
\description{
Retrieve sample financial data for analysis
}
\examples{
# Generate sample global markets data
data <- get_sample_data("global_markets", n_assets = 5, n_periods = 100)
head(data)

# Generate crypto data
crypto_data <- get_sample_data("crypto", n_assets = 3, n_periods = 50)

# Generate commodities data
commodity_data <- get_sample_data("commodities", n_assets = 4)
}
