% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{load_analysis_results}
\alias{load_analysis_results}
\title{Load Analysis Results}
\usage{
load_analysis_results(filename)
}
\arguments{
\item{filename}{Character string of file to load}
}
\value{
Loaded analysis results
}
\description{
Load previously saved analysis results
}
